Adding floating points to grammar

+++ tools/Assembler/atas.py
@@ -64,7 +64,7 @@ def assemble(f):
 			binary += "0001"
 		elif filter == "int":
 			binary += "0010"
-		elif filter == "double":
+		elif filter == "float":
 			binary += "0011"
 		elif filter == "rate":
 			binary += "0100"
